Updated CRC Capture Regex
It's now looking for 8 hexadecimal digits surrounded by either () or []
or actually because of how it works (] or [) are fine too but users have
no incentive to name their own files somethign that stupid. This should
avoid the problem I had with Kyolug Returns, but may not actually be any
better overall.