Added CRC Override flag
The regex to find the CRC in a filename has been simplified
significantly.

It also is no longer the only way to get a CRC into the metadata short
of manually opening the site and editing, as you can pass whatever you
want via the -c flag, and call it a day. I'm assuming nyaa doesn't
actually validate that these CRCs are even valid four byte hex
representations, so if the user wishes to feed garbage I'm not going to
stop them.