Fix Stupidity
So, that previous commit actually included everything for the feature...
except it called the wrong variable name. In fact, it called a variable
that didn't get populated to open as a file, accomplishing nothing.

By switching it to link_data_filename, which is y'know the string that
link_log writes to, it works. Thanks for backing up your submissions
long enough for me to test this feature, tosho!